refactor elasticsearch directory
