+++ modules/food/infras/repository/elasticsearch/delete_food.go
@@ -1,0 +1,12 @@
+package elasticsearch
+
+import (
+	"context"
+
+	"github.com/google/uuid"
+)
+
+// DeleteFood deletes a food document from Elasticsearch
+func (r *FoodSearchRepo) DeleteFood(ctx context.Context, id uuid.UUID) error {
+	return r.esClient.DeleteDocument(ctx, id.String())
+}
