+++ modules/restaurant/infras/repository/elasticsearch/initialize.go
@@ -1,0 +1,10 @@
+package elasticsearch
+
+import (
+	"context"
+)
+
+// Initialize creates the restaurant index with the proper mapping
+func (r *RestaurantSearchRepo) Initialize(ctx context.Context) error {
+	return r.esClient.CreateIndex(ctx, RestaurantIndexMapping)
+}
