+++ modules/restaurant/infras/repository/elasticsearch/reindex_all_restaurants.go
@@ -1,0 +1,14 @@
+package elasticsearch
+
+import (
+	"context"
+	"log"
+
+	restaurantmodel "github.com/ntttrang/go-food-delivery-backend-service/modules/restaurant/model"
+)
+
+// ReindexAllRestaurants reindexes all restaurants from the database
+func (r *RestaurantSearchRepo) ReindexAllRestaurants(ctx context.Context, restaurants []restaurantmodel.Restaurant) error {
+	log.Printf("Reindexing %d restaurants", len(restaurants))
+	return r.BulkIndexRestaurants(ctx, restaurants)
+}
