+++ modules/restaurant/module.go
@@ -1,12 +1,11 @@
 package restaurantmodule
 
 import (
-	"context"
 	"log"
 
 	"github.com/gin-gonic/gin"
 	restaurantHttpgin "github.com/ntttrang/go-food-delivery-backend-service/modules/restaurant/infras/controller/http-gin"
-	elasticsearch "github.com/ntttrang/go-food-delivery-backend-service/modules/restaurant/infras/repository/elasticsearch"
+	elasticsearchrepo "github.com/ntttrang/go-food-delivery-backend-service/modules/restaurant/infras/repository/elasticsearch"
 	restaurantgormmysql "github.com/ntttrang/go-food-delivery-backend-service/modules/restaurant/infras/repository/gorm-mysql"
 	rpcclient "github.com/ntttrang/go-food-delivery-backend-service/modules/restaurant/infras/repository/rpc-client"
 	restaurantService "github.com/ntttrang/go-food-delivery-backend-service/modules/restaurant/service"
@@ -25,6 +24,14 @@ func SetupRestaurantModule(appCtx shareinfras.IAppContext, g *gin.RouterGroup) {
 	restaurantLikeRepo := restaurantgormmysql.NewRestaurantLikeRepo(dbCtx)
 	restaurantRatingRepo := restaurantgormmysql.NewRestaurantRatingRepo(dbCtx)
 
+	userRPCClient := rpcclient.NewUserRPCClient(appCtx.GetConfig().UserServiceURL)
+
+	esClient, err := shareComponent.NewElasticsearchClient(appCtx.GetConfig().ElasticSearch)
+	if err != nil {
+		log.Printf("Elasticsearch initialization failed: %v. Search functionality will be disabled.", err)
+	}
+	esRestaurantRepo := elasticsearchrepo.NewRestaurantSearchRepo(esClient)
+
 	// Setup basic handlers
 	createCmdHdl := restaurantService.NewCreateCommandHandler(restaurantRepo, restaurantFoodRepo)
 	listQueryHdl := restaurantService.NewListQueryHandler(restaurantRepo, restaurantFoodRepo)
@@ -32,11 +39,11 @@ func SetupRestaurantModule(appCtx shareinfras.IAppContext, g *gin.RouterGroup) {
 	updateCmdHdl := restaurantService.NewUpdateRestaurantCommandHandler(restaurantRepo)
 	deleteCmdHdl := restaurantService.NewDeleteCommandHandler(restaurantRepo)
 
+	// Favorite restaurant
 	createRestaurantFavoriteCmdl := restaurantService.NewAddFavoritesCommandHandler(restaurantLikeRepo)
 	favoriteRestaurantQueryHdl := restaurantService.NewGetFavoritesRestaurantQueryHandler(restaurantRepo)
 
-	userRPCClient := rpcclient.NewUserRPCClient(appCtx.GetConfig().UserServiceURL)
-
+	// restaurant comment
 	createCommentRestaurantCmdl := restaurantService.NewCommentRestaurantCommandHandler(restaurantRatingRepo)
 	listCommentRestaurantCmdl := restaurantService.NewListRestaurantCommentsQueryHandler(restaurantRatingRepo, userRPCClient)
 	deleteCommentRestaurantCmdl := restaurantService.NewDeleteCommentCommandHandler(restaurantRatingRepo)
@@ -45,63 +52,17 @@ func SetupRestaurantModule(appCtx shareinfras.IAppContext, g *gin.RouterGroup) {
 	listMenuItemCmdHdl := restaurantService.NewListMenuItemQueryHandler(restaurantFoodRepo, foodRPCClient, catRPCClient)
 	deleteMenuItemCmdHdl := restaurantService.NewDeleteMenuItemCommandHandler(restaurantFoodRepo)
 
-	// Setup Elasticsearch if available
-	var searchRestaurantQueryHandler *restaurantService.SearchRestaurantQueryHandler
-	var syncRestaurantIndexCommandHandler *restaurantService.SyncRestaurantIndexCommandHandler
-
-	// Try to initialize Elasticsearch client
-	esClient, err := shareComponent.NewElasticsearchClient(appCtx.GetConfig().ElasticSearch)
-	if err != nil {
-		log.Printf("Elasticsearch initialization failed: %v. Search functionality will be disabled.", err)
-	}
-
-	// If Elasticsearch client was successfully created, set up search functionality
-	if esClient != nil {
-		// Create a new client with the restaurant index name
-		restaurantIndexName := "restaurants"
-
-		// Use the client with the restaurant index
-		restaurantEsClient := esClient.WithIndex(restaurantIndexName)
-
-		// Setup search repository
-		restaurantSearchRepo := elasticsearch.NewRestaurantSearchRepo(restaurantEsClient)
-
-		// Initialize the index with proper mapping
-		if err := restaurantSearchRepo.Initialize(context.Background()); err != nil {
-			log.Printf("Failed to initialize restaurant index: %v", err)
-		}
-
-		// Setup search handlers
-		searchRestaurantQueryHandler = restaurantService.NewSearchRestaurantQueryHandler(restaurantSearchRepo)
-		syncRestaurantIndexCommandHandler = restaurantService.NewSyncRestaurantIndexCommandHandler(restaurantRepo, restaurantSearchRepo)
-
-		// Setup event handler for Elasticsearch operations
-		// This handler would be used to hook into restaurant CRUD operations
-		// to automatically update the Elasticsearch index
-		_ = restaurantService.NewRestaurantElasticsearchHandler(restaurantSearchRepo)
-
-		log.Println("Elasticsearch initialized successfully. Restaurant search functionality is enabled.")
-	} else {
-		log.Println("Elasticsearch client not available. Restaurant search functionality will be disabled.")
-	}
-
-	// Create dummy handlers if Elasticsearch is not available
-	if searchRestaurantQueryHandler == nil {
-		// Create a dummy search handler that returns empty results
-		searchRestaurantQueryHandler = restaurantService.NewSearchRestaurantQueryHandler(nil)
-	}
-
-	if syncRestaurantIndexCommandHandler == nil {
-		// Create a dummy sync handler with nil repositories
-		syncRestaurantIndexCommandHandler = restaurantService.NewSyncRestaurantIndexCommandHandler(nil, nil)
-	}
+	// Setup Elasticsearch
+	searchRestaurantQueryHandler := restaurantService.NewSearchRestaurantQueryHandler(esRestaurantRepo)
+	syncRestaurantByIdCommandHandler := restaurantService.NewSyncRestaurantByIdCommandHandler(restaurantRepo, esRestaurantRepo)
+	syncRestaurantIndexCommandHandler := restaurantService.NewSyncRestaurantIndexCommandHandler(restaurantRepo, esRestaurantRepo)
 
 	resCtl := restaurantHttpgin.NewRestaurantHttpController(
 		createCmdHdl, listQueryHdl, getDetailQueryHdl, updateCmdHdl, deleteCmdHdl,
 		createRestaurantFavoriteCmdl, favoriteRestaurantQueryHdl,
 		createCommentRestaurantCmdl, listCommentRestaurantCmdl, deleteCommentRestaurantCmdl,
 		createMenuItemCmdHdl, listMenuItemCmdHdl, deleteMenuItemCmdHdl,
-		searchRestaurantQueryHandler, syncRestaurantIndexCommandHandler,
+		searchRestaurantQueryHandler, syncRestaurantByIdCommandHandler, syncRestaurantIndexCommandHandler,
 	)
 
 	restaurants := g.Group("/restaurants")
