+++ modules/restaurant/service/create_restaurant.go
@@ -33,38 +33,6 @@ func NewCreateCommandHandler(createRestaurantRepo ICreateRestaurantRepository, b
 	}
 }
 
-// func (s *CreateCommandHandler) Execute(ctx context.Context, req *restaurantmodel.RestaurantInsertDto) error {
-// 	if err := req.Validate(); err != nil {
-// 		return err
-// 	}
-
-// 	restaurant := req.ConvertToRestaurant()
-// 	restaurant.Id, _ = uuid.NewV7()
-// 	restaurant.Status = sharedModel.StatusActive // Always set Active Status when insert
-
-// 	if err := s.createRestaurantRepo.Insert(ctx, *restaurant); err != nil {
-// 		return err
-// 	}
-
-// 	foods := req.Foods
-// 	if len(foods) > 0 {
-// 		var restaurantFoods []restaurantmodel.RestaurantFood
-// 		for _, f := range foods {
-// 			restaurantFood := f.ConvertToRestaurantFood()
-// 			restaurantFood.RestaurantID = restaurant.Id
-// 			restaurantFoods = append(restaurantFoods, *restaurantFood)
-// 		}
-// 		if err := s.bulkRestaurantFoodRepo.BulkInsert(ctx, restaurantFoods); err != nil {
-// 			return err
-// 		}
-// 	}
-
-// 	// set data to response
-// 	req.Id = restaurant.Id
-
-// 	return nil
-// }
-
 func (s *CreateCommandHandler) Execute(ctx context.Context, req *restaurantmodel.RestaurantInsertDto) error {
 	if err := req.Validate(); err != nil {
 		return datatype.ErrBadRequest.WithWrap(err).WithDebug(err.Error())
