+++ modules/restaurant/service/sync_restaurant_index.go
@@ -4,20 +4,17 @@ import (
 	"context"
 	"log"
 
-	"github.com/google/uuid"
 	restaurantmodel "github.com/ntttrang/go-food-delivery-backend-service/modules/restaurant/model"
 	"github.com/ntttrang/go-food-delivery-backend-service/shared/datatype"
 )
 
 // IRestaurantRepo defines the interface for restaurant repository operations
 type IRestaurantRepo interface {
-	FindById(ctx context.Context, id uuid.UUID) (*restaurantmodel.Restaurant, error)
 	FindAll(ctx context.Context) ([]restaurantmodel.Restaurant, error)
 }
 
 // IRestaurantIndexRepo defines the interface for restaurant indexing operations
 type IRestaurantBulkIndexRepo interface {
-	IndexRestaurant(ctx context.Context, restaurant *restaurantmodel.Restaurant) error
 	ReindexAllRestaurants(ctx context.Context, restaurants []restaurantmodel.Restaurant) error
 }
 
@@ -40,11 +37,6 @@ func NewSyncRestaurantIndexCommandHandler(
 
 // SyncAll synchronizes all restaurants with Elasticsearch
 func (s *SyncRestaurantIndexCommandHandler) SyncAll(ctx context.Context) error {
-	// Check if repositories are available
-	if s.indexRepo == nil || s.restaurantRepo == nil {
-		return datatype.ErrInternalServerError.WithDebug("Elasticsearch functionality is not available. Elasticsearch is not configured.")
-	}
-
 	// Get all restaurants from the database
 	restaurants, err := s.restaurantRepo.FindAll(ctx)
 	if err != nil {
@@ -55,25 +47,6 @@ func (s *SyncRestaurantIndexCommandHandler) SyncAll(ctx context.Context) error {
 
 	// Reindex all restaurants
 	if err := s.indexRepo.ReindexAllRestaurants(ctx, restaurants); err != nil {
-		return datatype.ErrInternalServerError.WithWrap(err).WithDebug(err.Error())
-	}
-
-	return nil
-}
-
-// SyncRestaurant synchronizes a single restaurant with Elasticsearch
-func (s *SyncRestaurantIndexCommandHandler) SyncRestaurant(ctx context.Context, id uuid.UUID) error {
-	// Check if repositories are available
-	if s.indexRepo == nil || s.restaurantRepo == nil {
-		return datatype.ErrInternalServerError.WithDebug("Elasticsearch functionality is not available. Elasticsearch is not configured.")
-	}
-
-	restaurant, err := s.restaurantRepo.FindById(ctx, id)
-	if err != nil {
-		return datatype.ErrInternalServerError.WithWrap(err).WithDebug(err.Error())
-	}
-
-	if err := s.indexRepo.IndexRestaurant(ctx, restaurant); err != nil {
 		return datatype.ErrInternalServerError.WithWrap(err).WithDebug(err.Error())
 	}
